Moved Index File Within lib Folder + Updated Return Types for Lists and People
Moved index.js into the lib folder. Updated all references to the index file. index.js comments were also updated with return types and more through comments.

Updated the return types for lists and people in preparation of new documentation.
