+++ lib/index.js
@@ -1,10 +1,14 @@
-const User = require("./lib/user"),
-    lists = require("./lib/lists"),
-    media = require("./lib/media"),
-    people = require("./lib/people"),
-    util = require("./lib/utilities");
+const User = require("./user"),
+    lists = require("./lists"),
+    media = require("./media"),
+    people = require("./people"),
+    util = require("./utilities");
 
-module.exports = class AniList {
+/**
+ * The main class for AniList-Node
+ * @since 1.0.0
+ */
+class AniList {
     /**
      * @constructor
      * @param {String} [accessKey] - The AniList API token. If no key is provided, 
@@ -17,19 +21,35 @@ module.exports = class AniList {
         // Import utilites for the classes.
         this.__util = new util(accessKey);
 
+        /**
+         * @Member { User }
+         */
         this.user = new User(this.__util);
+
+        /**
+         * @Member { Lists }
+         */
         this.lists = new lists(this.__util);
+
+        /**
+         * @Member { Media }
+         */
         this.media = new media(this.__util);
+
+        /**
+         * @Member { People }
+         */
         this.people = new people(this.__util);
     };
 
     /**
      * Grabs data on a studio
-     * @param {String | Number} id - The studio ID or name on AniList.
-     * @returns { Object } Returns a customized data object.
+     * @param {String | Number} studio - The studio ID or name on AniList.
+     * @returns { StudioEntry }
+     * @since 1.0.0
      */
-    studio(id) {
-        var queryVars = this.__util.generateQueryHeaders("Studio", id);
+    studio(studio) {
+        var queryVars = this.__util.generateQueryHeaders("Studio", studio);
 
         return this.__util.send(queryVars[1] + `id name isAnimationStudio siteUrl isFavourite favourites 
             media { nodes { id title { romaji english native userPreferred } } } } }`, queryVars[0]);
@@ -37,10 +57,12 @@ module.exports = class AniList {
 
     /**
      * Searchs AniList based on a specific term.
-     * @param {String} type - Required. Either anime, manga, character, staff or studio. 
-     * @param {String} term - Required. The term to lookup.
+     * @param {String} type - Required. Either anime, manga, character, staff, studio, or user. 
+     * @param {String} term - Required. The term to lookup. (ie: "Honzuki no Gekokujou" or "Butterstroke")
      * @param {Number} page - Which page of the results to look at. Will default to 1 if not provided.
      * @param {Number} amount - The amount of results per page. AniList will cap this at 25 and function will default to 5 if not provided.
+     * @return { SearchEntry }
+     * @since 1.0.0
      */
     search(type, term, page=1, amount=5) {
         if (!type) { throw new Error("Type of search not defined!"); }
@@ -57,19 +79,16 @@ module.exports = class AniList {
             "manga": "media (type: MANGA, search: $search) { id title { romaji english native userPreferred } }",
             "char": "characters (search: $search) { id name { english: full } }" ,
             "staff": "staff (search: $search) { id name { english: full } }",
-            "studio": "studios (search: $search) { id name }"
-        }
+            "studio": "studios (search: $search) { id name }",
+            "user": "users (search: $search) { id name }"
+        };
 
-        switch (type.toLowerCase()) {
-            case "anime": var query = search["anime"]; break;
-            case "manga": var query = search["manga"]; break;
-            case "character": var query = search["char"]; break;
-            case "staff": var query = search["staff"]; break;
-            case "studio": var query = search["studio"]; break;
-            default: throw new Error("Type not supported.");
-        }
+        var query = search[type.toLowerCase()];
+        if (!query) { throw new Error("Type not supported."); }
 		
         return this.__util.send(`query ($page: Int, $perPage: Int, $search: String) {
         Page (page: $page, perPage: $perPage) { pageInfo { total currentPage lastPage hasNextPage perPage } ${query} } }`, { search: term, page: page, perPage: amount});
     };
 };
+
+module.exports = AniList;
