+++ lib/lists.js
@@ -16,7 +16,7 @@ class Lists {
     /**
      * Fetch a user's AniList anime lists.
      * @param { Number | String } user - Required. Can either be the username or the AniList ID.
-     * @returns { Object } Returns a customized data object.
+     * @returns { UserList }
      * @since 1.1.0
      */
     anime(user) {
@@ -33,7 +33,7 @@ class Lists {
     /**
      * Fetch a user's AniList manga lists.
      * @param { Number | String } user - Required. Can either be the username or the AniList ID.
-     * @returns { Object } Returns a customized data object.
+     * @returns { UserList }
      * @since 1.1.0
      */
     manga(user) {
