+++ lib/people.js
@@ -16,7 +16,7 @@ class People {
     /**
      * Fetch a character entry by its AniList ID.
      * @param { Number|String } id - Required. The ID tied to the AniList entry.
-     * @returns { Object } Returns a customized data object.
+     * @returns { CharacterEntry }
      * @since 1.0.0
      */
     character(id) {
@@ -30,7 +30,7 @@ class People {
     /**
      * Fetch a staff entry by its AniList ID or their name.
      * @param { Number|String } id - Required. The ID can either be the AniList ID or the staff's name.
-     * @returns { Object } Returns a customized data object.
+     * @returns { StaffEntry }
      * @since 1.0.0
      */
     staff(id) {
