+++ package.json
@@ -1,10 +1,11 @@
 {
   "name": "anilist-node",
-  "version": "1.5.0",
+  "version": "1.6.0",
   "description": "A lightweight Node.js wrapper for the AniList API",
-  "main": "index.js",
+  "main": "./lib/index.js",
   "scripts": {
-    "test": "node tests/mediaTest.js"
+    "test": "node tests/mediaTest.js",
+    "docs": "jsdoc -c docConfig.json"
   },
   "author": {
     "name": "Frederick Katsura",
@@ -26,5 +27,8 @@
     "api",
     "anime",
     "node.js"
-  ]
+  ],
+  "devDependencies": {
+    "docdash": "^1.2.0"
+  }
 }
