+++ tests/searchTest.js
@@ -1,4 +1,4 @@
-const anilistnode = require("../index");
+const anilistnode = require("../lib/index");
 const Anilist = new anilistnode();
 
 Anilist.search("anime", "Occultic;Nine", 1, 10).then((data) => { console.log(data); });
